refactor: 그룹 참여 / 생성
