+++ apps/mac/src/renderer/src/features/record/ui/group-side-tab/GroupSideTab.tsx
@@ -3,12 +3,15 @@ import * as S from "./GroupSideTab.style";
 import { type Group as GroupEntity, GROUP_CATEGORY_LABELS } from "@/entities/group";
 import { useGetMyProfile } from "@/entities/user";
 import { Popover } from "@/shared/ui";
+import type { GroupFormPanelProps } from "../../model/useGroup";
+import { GroupFormPanel } from "./panel/GroupFormPanel";
 
 interface GroupSideTabProps {
   groups: GroupEntity[];
   currentGroupId: number | null;
   onSelectGroup: (groupId: number) => void;
-  onOpenFormModal: () => void;
+  onOpenFormPanel: () => void;
+  formPanel: GroupFormPanelProps;
   onEditGroup: (groupId: number) => void;
   onDeleteGroup: (groupId: number) => void;
   onQuitGroup: (groupId: number) => void;
@@ -18,7 +21,8 @@ export const GroupSideTab = ({
   groups,
   currentGroupId,
   onSelectGroup,
-  onOpenFormModal,
+  onOpenFormPanel,
+  formPanel,
   onEditGroup,
   onDeleteGroup,
   onQuitGroup,
@@ -38,83 +42,102 @@ export const GroupSideTab = ({
 
   return (
     <S.GroupSideTabContainer>
-      <S.Title>그룹 목록</S.Title>
-      <S.GroupList>
-        {groups.map(group => {
-          const isSelected = group.id === currentGroupId;
-          const isOwner = myUserId === group.owner.id;
-          const isMenuOpen = openMenuGroupId === group.id;
+      {formPanel.isOpen ? (
+        <S.FormPanelWrapper>
+          <GroupFormPanel {...formPanel} />
+        </S.FormPanelWrapper>
+      ) : (
+        <>
+          <S.Title>그룹 목록</S.Title>
+          <S.GroupList>
+            {groups.map(group => {
+              const isSelected = group.id === currentGroupId;
+              const isOwner = myUserId === group.owner.id;
+              const isMenuOpen = openMenuGroupId === group.id;
 
-          return (
-            <S.GroupListItem key={group.id} $isSelected={isSelected}>
-              <S.GroupSelectButton type="button" onClick={() => onSelectGroup(group.id)}>
-                <S.GroupListLeft>
-                  <S.SelectionDot $isSelected={isSelected} />
-                  <S.GroupListName>{group.name}</S.GroupListName>
-                </S.GroupListLeft>
-              </S.GroupSelectButton>
-              <S.GroupListRight>
-                <S.GroupCategory>{GROUP_CATEGORY_LABELS[group.category]}</S.GroupCategory>
-                <S.MoreIconWrapper ref={isMenuOpen ? menuRef : null}>
-                  <S.IconButton
-                    type="button"
-                    aria-label="그룹 메뉴 열기"
-                    onClick={event => {
-                      event.stopPropagation();
-                      handleMoreClick(group.id);
-                    }}
-                  >
-                    <S.MoreIcon />
-                  </S.IconButton>
-                  <Popover
-                    isOpen={isMenuOpen}
-                    onClose={handleCloseMenu}
-                    anchorRef={menuRef}
-                    minWidth="max-content"
-                  >
-                    <S.MenuList>
-                      {isOwner ? (
-                        <>
-                          <S.MenuItem
-                            type="button"
-                            onClick={() => {
-                              handleCloseMenu();
-                              onEditGroup(group.id);
-                            }}
-                          >
-                            그룹 수정
-                          </S.MenuItem>
-                          <S.MenuItem
-                            type="button"
-                            onClick={() => {
-                              handleCloseMenu();
-                              onDeleteGroup(group.id);
-                            }}
-                          >
-                            그룹 삭제
-                          </S.MenuItem>
-                        </>
-                      ) : (
-                        <S.MenuItem
-                          type="button"
-                          onClick={() => {
-                            handleCloseMenu();
-                            onQuitGroup(group.id);
-                          }}
-                        >
-                          그룹 탈퇴
-                        </S.MenuItem>
-                      )}
-                    </S.MenuList>
-                  </Popover>
-                </S.MoreIconWrapper>
-              </S.GroupListRight>
-            </S.GroupListItem>
-          );
-        })}
-      </S.GroupList>
-      <S.PlusIconWrapper type="button" onClick={onOpenFormModal}>
-        <S.PlusIcon />
+              return (
+                <S.GroupListItem key={group.id} $isSelected={isSelected}>
+                  <S.GroupSelectButton type="button" onClick={() => onSelectGroup(group.id)}>
+                    <S.GroupListLeft>
+                      <S.SelectionDot $isSelected={isSelected} />
+                      <S.GroupListName>{group.name}</S.GroupListName>
+                    </S.GroupListLeft>
+                  </S.GroupSelectButton>
+                  <S.GroupListRight>
+                    <S.GroupCategory>{GROUP_CATEGORY_LABELS[group.category]}</S.GroupCategory>
+                    <S.MoreIconWrapper ref={isMenuOpen ? menuRef : null}>
+                      <S.IconButton
+                        type="button"
+                        aria-label="그룹 메뉴 열기"
+                        onClick={event => {
+                          event.stopPropagation();
+                          handleMoreClick(group.id);
+                        }}
+                      >
+                        <S.MoreIcon />
+                      </S.IconButton>
+                      <Popover
+                        isOpen={isMenuOpen}
+                        onClose={handleCloseMenu}
+                        anchorRef={menuRef}
+                        minWidth="max-content"
+                      >
+                        <S.MenuList>
+                          {isOwner ? (
+                            <>
+                              <S.MenuItem
+                                type="button"
+                                onClick={() => {
+                                  handleCloseMenu();
+                                  onEditGroup(group.id);
+                                }}
+                              >
+                                그룹 수정
+                              </S.MenuItem>
+                              <S.MenuItem
+                                type="button"
+                                onClick={() => {
+                                  handleCloseMenu();
+                                  onDeleteGroup(group.id);
+                                }}
+                              >
+                                그룹 삭제
+                              </S.MenuItem>
+                            </>
+                          ) : (
+                            <S.MenuItem
+                              type="button"
+                              onClick={() => {
+                                handleCloseMenu();
+                                onQuitGroup(group.id);
+                              }}
+                            >
+                              그룹 탈퇴
+                            </S.MenuItem>
+                          )}
+                        </S.MenuList>
+                      </Popover>
+                    </S.MoreIconWrapper>
+                  </S.GroupListRight>
+                </S.GroupListItem>
+              );
+            })}
+          </S.GroupList>
+        </>
+      )}
+      <S.PlusIconWrapper
+        type="button"
+        aria-label={formPanel.isOpen ? "그룹 패널 닫기" : "그룹 패널 열기"}
+        onClick={() => {
+          handleCloseMenu();
+          if (formPanel.isOpen) {
+            formPanel.onClose();
+            return;
+          }
+          onOpenFormPanel();
+        }}
+      >
+        <S.PlusIcon $isOpen={formPanel.isOpen} />
       </S.PlusIconWrapper>
     </S.GroupSideTabContainer>
   );
